load testing with k6

+++ app/src/main/java/org/dismefront/api/coordinates/CoordinatesController.java
@@ -3,6 +3,8 @@ package org.dismefront.api.coordinates;
 import lombok.RequiredArgsConstructor;
 import org.dismefront.data.coordinates.CoordinatesRepository;
 import org.dismefront.data.coordinates.CoordinatesService;
+import org.dismefront.data.event.EventName;
+import org.dismefront.data.event.EventRepository;
 import org.dismefront.data.shared.Role;
 import org.dismefront.data.user.User;
 import org.dismefront.data.user.UserRepository;
@@ -20,6 +22,7 @@ public class CoordinatesController {
     private final CoordinatesRepository coordinatesRepository;
     private final CoordinatesService coordinatesService;
     private final UserRepository userRepository;
+    private final EventRepository eventRepository;
 
     @GetMapping("/get-all")
     public ResponseEntity getAll() {
@@ -42,6 +45,10 @@ public class CoordinatesController {
     @PutMapping("/update/{id}")
     public ResponseEntity update(@PathVariable long id, @RequestBody CoordinatesRequest coordinatesRequest, Principal principal) {
         String username = principal.getName();
+        User user = userRepository.findByUsername(username).orElseThrow();
+        if (eventRepository.findByNameAndActorAndEntityId(EventName.COORDINATE_CREATED, username, id).isEmpty() && !user.getRoles().contains(Role.ROLE_ADMIN)) {
+            return ResponseEntity.status(410).body("You cannot delete this coordinate");
+        }
         try {
             if (coordinatesRequest.getY() == null) {
                 return ResponseEntity.badRequest().body("Y coordinate cannot be null");
@@ -55,6 +62,10 @@ public class CoordinatesController {
     @PostMapping("/delete/{id}")
     public ResponseEntity delete(@PathVariable long id, Principal principal) {
         String username = principal.getName();
+        User user = userRepository.findByUsername(username).orElseThrow();
+        if (eventRepository.findByNameAndActorAndEntityId(EventName.COORDINATE_CREATED, username, id).isEmpty() && !user.getRoles().contains(Role.ROLE_ADMIN)) {
+            return ResponseEntity.status(410).body("You cannot delete this coordinate");
+        }
         try {
             coordinatesService.deleteCoordinates(id, username);
             return ResponseEntity.ok().build();
